Refactor jisyo.ts
