+++ src/jisyo/types.ts
@@ -1,0 +1,4 @@
+import { Candidate } from "./candidate";
+
+export type Jisyo = Map<string, Candidate[]>;
+export type CacheMetadata = { expiry: number };
